Add 'count' flag to RedThing.find method and delete tmp_keys earlier
As the number of RedThing objects increases, the indexes will take up
more and more memory. The temporary sets and sorted sets that are
generated during different phases of 'find' could take up a decent
amount of space, so they should be deleted as soon as possible.

Also, update some asserts to use count flag in TestRedThing.test_find