Update RedThing.find to return a list (or int) instead of a generator
It is probably important that all the temporary sets and zsets get
removed from redis as quickly as possible. If some process were
iterating over the results in a slow way, it could be a while before the
final 'yield' occurs and a potentially massive set/zset could be taking
up space in redis (especially if there are many simultaneous find calls
happening).